+++ src/flags.py
@@ -24,10 +24,25 @@ __license__ = 'MIT'
 
 
 def unique(flags_class):
+    """ A decorator for flags classes to forbid flag aliases. """
     if not flags_class.__member_aliases__:
         return flags_class
     aliases = ', '.join('%s -> %s' % (alias, name) for alias, name in flags_class.__member_aliases__.items())
     raise ValueError('duplicate values found in %r: %s' % (flags_class, aliases))
+
+
+def unique_bits(flags_class):
+    """ A decorator for flags classes to forbid declaring flags with overlapping bits. """
+    flags_class = unique(flags_class)
+    other_bits = 0
+    for name, member in flags_class.__members_without_aliases__.items():
+        bits = int(member)
+        if other_bits & bits:
+            for other_name, other_member in flags_class.__members_without_aliases__.items():
+                if int(other_member) & bits:
+                    raise ValueError("%r: '%s' and '%s' have overlapping bits" % (flags_class, other_name, name))
+        else:
+            other_bits |= bits
 
 
 def _is_descriptor(obj):