+++ CHANGES.rst
@@ -8,7 +8,9 @@ Future version
 
 *Release date: future*
 
-- Added the ``@unique`` decorator that can be used to forbid aliases in a flags class.
+- Added the ``@unique_bits`` decorator that can be used to forbid overlapping bits between the values of flags.
+- Added the ``@unique`` decorator that can be used to forbid aliases in a flags class. (An alias is a flag that
+  has the exact same value as a previously declared flag.)
 - Added ``__repr__`` for flags classes.